project/requirements.txt
fastapi==0.60.1
pandas==1.1.0
plotly==4.9.0
uvicorn==0.11.8
psycopg2-binary==2.8.5
SQLAlchemy==1.3.19
python-dotenv==0.14.0